fix(T1408): sanitize flag reason + PWA icons + Prometheus alerts + projects split
- Sanitize `reason` field in PATCH /api/tasks/:id/flag (stored XSS fix)
- Add PWA icons (192x192, 512x512) for manifest.json
- Add Prometheus alert rules for titan-app (error rate, latency, memory)
- Split projects/page.tsx (2400+ lines) into 3 focused components

Closes #1409
